+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this library are documented in this file.
 
 - Better handling of Twitter/X status 403 errors (#156).
 - Improved atmosphere test coverage and added `pytest-timeout` dev dep (#159).
+- Improved ATmosphere handling of a common `InvokeTimeoutError` (#168).
 
 ### Bug Fixes